transform-sdk/go: move to transform directory
Go uses the repo layout as the package name, so don't let our monorepo
structure mislead users to what the sdk's package name is.

An upcoming commit will rename the package to `transform`.
