+++ .github/workflows/transform-sdk-build.yml
@@ -39,7 +39,7 @@ jobs:
           go-version: 1.20.6
 
       - name: Run tests
-        working-directory: src/transform-sdk/go
+        working-directory: src/transform-sdk/go/transform
         run: go test -v ./...
 
       - name: Run golangci-lint
@@ -47,4 +47,4 @@ jobs:
         with:
           version: latest
           args: --timeout 5m
-          working-directory: src/transform-sdk/go
+          working-directory: src/transform-sdk/go/transform
