+++ src/CMakeLists.txt
@@ -17,4 +17,4 @@ find_package(absl REQUIRED)
 find_program(GO_PROGRAM go REQUIRED)
 add_subdirectory(v)
 add_subdirectory(go/kreq-gen)
-add_subdirectory(transform-sdk/go/internal/testdata)
+add_subdirectory(transform-sdk/go/transform/internal/testdata)
