+++ src/go/rpk/pkg/cli/transform/template/golang/transform.gosrc
@@ -1,7 +1,7 @@
 package main
 
 import (
-	"github.com/redpanda-data/redpanda/src/transform-sdk/go"
+	"github.com/redpanda-data/redpanda/src/transform-sdk/go/transform"
 )
 
 func main() {
