--- src/transform-sdk/go/go.mod
@@ -1,3 +1,0 @@
-module github.com/redpanda-data/redpanda/src/transform-sdk/go
-
-go 1.20
