--- src/transform-sdk/go/internal/testdata/go.mod
@@ -1,10 +1,0 @@
-module github.com/redpanda-data/wasm-transform-testdata
-
-go 1.20
-
-require (
-	github.com/actgardner/gogen-avro/v10 v10.2.1
-	github.com/redpanda-data/redpanda/src/transform-sdk/go v0.0.1
-)
-
-replace github.com/redpanda-data/redpanda/src/transform-sdk/go => ../../
