+++ src/transform-sdk/go/transform/example_mirror_test.go
@@ -15,7 +15,7 @@
 package redpanda_test
 
 import (
-	"github.com/redpanda-data/redpanda/src/transform-sdk/go"
+	"github.com/redpanda-data/redpanda/src/transform-sdk/go/transform"
 )
 
 // This example shows the basic usage of the package:
