+++ src/transform-sdk/go/transform/go.mod
@@ -1,0 +1,3 @@
+module github.com/redpanda-data/redpanda/src/transform-sdk/go/transform
+
+go 1.20
