+++ src/transform-sdk/go/transform/internal/testdata/build_all.sh
