+++ src/transform-sdk/go/transform/internal/testdata/schema-registry/avro/schema.avsc
