+++ src/transform-sdk/go/transform/internal/testdata/setup-panic/transform.go
