+++ src/transform-sdk/go/transform/internal/testdata/transform-panic/transform.go
@@ -15,35 +15,13 @@
 package main
 
 import (
-	"encoding/json"
-	"math/rand"
-	"os"
-	"time"
-
-	"github.com/redpanda-data/redpanda/src/transform-sdk/go"
+	"github.com/redpanda-data/redpanda/src/transform-sdk/go/transform"
 )
 
 func main() {
-	redpanda.OnRecordWritten(wasiTransform)
+	redpanda.OnRecordWritten(panicTransform)
 }
 
-type WasiInfo struct {
-	Args         []string
-	Env          []string
-	NowNanos     int64
-	RandomNumber int
-}
-
-func wasiTransform(e redpanda.WriteEvent) ([]redpanda.Record, error) {
-	w := &WasiInfo{
-		Args:         os.Args,
-		Env:          os.Environ(),
-		NowNanos:     time.Now().UnixNano(),
-		RandomNumber: rand.Int(),
-	}
-	b, err := json.Marshal(w)
-	if err != nil {
-		return nil, err
-	}
-	return []redpanda.Record{{Value: b}}, nil
+func panicTransform(e redpanda.WriteEvent) ([]redpanda.Record, error) {
+	panic("oh noes!")
 }
