+++ src/transform-sdk/go/transform/processor.go
@@ -19,7 +19,7 @@ import (
 	"time"
 	"unsafe"
 
-	"github.com/redpanda-data/redpanda/src/transform-sdk/go/internal/rwbuf"
+	"github.com/redpanda-data/redpanda/src/transform-sdk/go/transform/internal/rwbuf"
 )
 
 type batchHeader struct {
