+++ src/transform-sdk/go/transform/serialize.go
@@ -17,7 +17,7 @@ package redpanda
 import (
 	"encoding/binary"
 
-	"github.com/redpanda-data/redpanda/src/transform-sdk/go/internal/rwbuf"
+	"github.com/redpanda-data/redpanda/src/transform-sdk/go/transform/internal/rwbuf"
 )
 
 // Reusable slice of record headers
