+++ src/transform-sdk/go/transform/sr/encoding_test.go
@@ -18,7 +18,7 @@ import (
 	"reflect"
 	"testing"
 
-	"github.com/redpanda-data/redpanda/src/transform-sdk/go/internal/rwbuf"
+	"github.com/redpanda-data/redpanda/src/transform-sdk/go/transform/internal/rwbuf"
 )
 
 func TestSchemaRoundtrip(t *testing.T) {
