+++ src/transform-sdk/go/transform/sr/stub_abi.go
