+++ src/transform-sdk/go/transform/stub_abi.go
