+++ tests/docker/ducktape-deps/tinygo-wasi-transforms
@@ -4,7 +4,7 @@ set -x
 
 mkdir -p /opt/transforms/tinygo/
 
-cd /transform-sdk/go/internal/testdata
+cd /transform-sdk/go/transform/internal/testdata
 
 tinygo build -target wasi -opt=z \
   -panic print -scheduler none \
